RefixMatch/refix match image dockerized web app/requirements.txt
flask
Pillow
numpy<2
torch==2.1.0+cpu
torchvision==0.16.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
